Add files via upload
The data used for the choropleth is transferred into/out of Mongo DB and is hosted by the Flask API. The app.py is the flask app build and the Mongo DB Export.json file is the mongo file. 

The PPT presentation is also attached. 

Please initialize the website with the index.html